chore: fix eslint configuration
