+++ metro.config.js
@@ -1,10 +1,3 @@
-/**
- * Metro configuration for React Native
- * https://github.com/facebook/react-native
- *
- * @format
- */
-
 module.exports = {
   transformer: {
     getTransformOptions: async () => ({
@@ -14,4 +7,4 @@ module.exports = {
       },
     }),
   },
-};
+}
