+++ package.json
@@ -29,8 +29,13 @@
     "@typescript-eslint/parser": "^5.32.0",
     "babel-jest": "^28.1.3",
     "eslint": "^8.21.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.30.1",
+    "eslint-plugin-react-native": "^4.0.0",
     "jest": "^28.1.3",
     "metro-react-native-babel-preset": "^0.72.0",
+    "prettier": "^2.7.1",
     "react-test-renderer": "18.2.0",
     "typescript": "^4.4.4"
   },
